Split DefaultScalarGenerator enum into per-type generators
Replace the DefaultScalarGenerator enum with five standalone structs —
BooleanGenerator, IntGenerator, FloatGenerator, StringGenerator, IdGenerator —
each implementing the Generator trait directly. Move boxed() to a default
method on the trait, and collapse the generators/ directory into a single
generators.rs.
