+++ crates/apollo-smith/src/lib.rs
@@ -49,9 +49,13 @@ pub use enum_::EnumTypeDef;
 use field::FieldDef;
 pub use fragment::FragmentDef;
 pub use generators::default_generators;
-pub use generators::DefaultScalarGenerator;
+pub use generators::BooleanGenerator;
+pub use generators::FloatGenerator;
 pub use generators::Generator;
 pub use generators::Generators;
+pub use generators::IdGenerator;
+pub use generators::IntGenerator;
+pub use generators::StringGenerator;
 pub use input_object::InputObjectTypeDef;
 pub use interface::InterfaceTypeDef;
 use name::Name;
